Terror, horror, pain, suffering

+++ include/policy/first_fit.h
@@ -5,7 +5,7 @@
 class first_fit : public policy
 {
 public:
-    virtual pair<Mem_manager::book *, int> find_block(char *p, size_t size, size_t req);
+    virtual pair<Mem_manager::book *, size_t> find_block(char *p, size_t size, size_t req);
 };
 
 #endif
