+++ include/policy/policy.h
@@ -2,14 +2,14 @@
 #define POLICY_H
 #include <utility>
 #include <climits>
-#include "../mm/mm.h"
+#include "../meta/meta.h"
 using namespace std;
 
 class policy
 {
 public:
     virtual ~policy(){};
-    virtual pair<Mem_manager::book* , int> find_block(char *p, size_t size, size_t req) = 0;
+    virtual pair<Mem_manager::book* , size_t> find_block(char *p, size_t size, size_t req) = 0;
 };
