+++ src/policy/best_fit.cpp
@@ -1,9 +1,9 @@
 #include "../../include/policy/best_fit.h"
 
-pair<Mem_manager::book *, int> best_fit::find_block(char *p, size_t size, size_t req)
+pair<Mem_manager::book *, size_t> best_fit::find_block(char *p, size_t size, size_t req)
 {
-    int chunk_size;
-    int found_size = INT_MAX;
+    size_t chunk_size;
+    size_t found_size = INT_MAX;
     Mem_manager::book *target;
     Mem_manager::book *pres = (Mem_manager::book *)p; //points to the first Mem_manager::book
     while (pres != (Mem_manager::book *)(p + size))
